API grid calls

+++ CricektApp.Domain/PlayerPastRecord.cs
@@ -26,7 +26,7 @@ namespace CricketApp.Domain
         public int? NumberOf50s { get; set; }
         public int? NumberOf100s { get; set; }
         public int? BestScore { get; set; }
-        public int? TotalOvers { get; set; }
+        public float? TotalOvers { get; set; }
         public int? TotalBallRuns { get; set; }
         public int? TotalWickets { get; set; }
         public int? TotalMaidens { get; set; }
